requirements.txt
numpy==1.22.3
pandas==1.4.1
pandas_datareader==0.10.0
pymongo==4.0.2
python_dateutil==2.8.2
requests==2.27.1
setuptools==45.2.0
six==1.14.0
